requirements.txt
fastapi
uvicorn
playwright
openai
supabase
pydantic
pydantic-settings
python-multipart
httpx
jinja2
beautifulsoup4
pytest
groq
stripe
python-jose[cryptography]
passlib[bcrypt]
google-generativeai
apscheduler
feedparser
